splited to local and global useColor logic
